requirements.txt

python-telegram-bot
openai
google-cloud-vision
python-dotenv
pandas
fastapi
uvicorn[standard]
gunicorn
pytest
pytest-asyncio